+++ CHANGELOG.md
@@ -11,3 +11,8 @@ Auth plugin 30-06-2014
 ----------------------
 
  * Replaced depricated functions and made plugin ready for Moodle 2.7. 
+
+Auth plugin 02-06-2014
+----------------------
+
+ * Add the possibility to sync users via another module